changed variable 'logfile' to 'tracefile' in the server; the path to the tracefile is now set via cmd arguments; if this path does not exist, an error message is printed and the peer-process exits immediately

+++ peers_seperate/peer.py
@@ -7,19 +7,20 @@
 from client import Client
 from server import Server
 from logging.logger import Logger
+import os
 import argparse
 
 class Peer:
 
-    def __init__(self, hostport, others, coordinates, logDir):
+    def __init__(self, hostport, others, coordinates, tracefile, logDir):
         # 1 publisher
         # n subscribers
         self.logger = Logger(logDir);
         self.clients = []
-        self.server = Server(hostport, coordinates, self)
-        self.addresses = others.split(', ')
+        self.server = Server(hostport, coordinates, tracefile, self)
+        self.addresses = others.split(',')
         for i in range(len(self.addresses)):
-            self.clients.append(Client(hostport, self.addresses[i], self))
+            self.clients.append(Client(hostport, self.addresses[i].strip(), self))
 
         self.server.start()
         for i in range(len(self.clients)):
@@ -42,10 +43,14 @@ if __name__ == "__main__":
                         help='the port number of the server')
 
     parser.add_argument('--clients', type=str,
-                        help='IP-address and port of the clients')
+                        help='IP-address and port of the clients; \nExample: "localhost:5557,localhost:5558,localhost:5559"')
 
     parser.add_argument('--coordinates', type=str,
-                        help='coordinates of the left upper corner and the right lower corner for the rectangle which the server should observe')
+                        help='coordinates of the left upper corner and the right lower corner for the rectangle '
+                             'which the server should observe; \nExample: "0,65000,65000,0"')
+
+    parser.add_argument('--tracefile', type=str,
+                        help='the path to the tracefile which the server should read from')
 
     parser.add_argument('--logDir', type=str,
                         help='the directory in which the logfile should be saved')
@@ -55,9 +60,13 @@ if __name__ == "__main__":
     # --serverPort=5556
     # --addresses="localhost:5557, localhost:5558, localhost:5559"
     #log the params in file
-    paramsLog = open("paramsLog.txt", "w");
-    paramsLog.write("--serverPort=%d\n--clients=%s\n--coordinates=%s\n--logDir=%s"
+
+    if(not(os.path.exists(args.tracefile))):
+        print("Path to tracefile does not exist!\n")
+    else:
+        paramsLog = open("paramsLog.txt", "w");
+        paramsLog.write("--serverPort=%d\n--clients=%s\n--coordinates=%s\n--logDir=%s"
                     %(args.serverPort, args.clients, args.coordinates, args.logDir));
-    paramsLog.close()
-    Peer(args.serverPort, args.clients, args.coordinates, args.logDir)
+        paramsLog.close()
+        Peer(args.serverPort, args.clients, args.coordinates, args.tracefile, args.logDir)
